requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-environ==0.11.2
cryptography==42.0.8
django-axes==6.5.2
pytest==8.3.5
pytest-django==4.10.0